Updated site config file template's default API
So as to match the change in Staticman's default API introduced in #36.  Note
that #36 is a breaking change. for users who have set Staticman `endpoint` to
the empty string "".  However, the official production instance has already
broken, as reported in eduardoboucas/staticman#307, so this commit is simply an
info update.